chore: amino test helpers file, fix boolean handling on BuyDirect
