+++ packages/api/e2e/amino.spec.ts
@@ -21,9 +21,12 @@ import {
   MsgCancelSellOrder,
 } from '../src/generated/regen/ecocredit/marketplace/v1/tx';
 import Long from 'long';
-import * as crypto from 'crypto';
-import * as ethers from 'ethers';
 import { MessageClient } from '../src/tx/msg';
+import {
+  genRandomStr,
+  makeEthContract,
+  makeEthTxHash,
+} from './amino-spec-helpers';
 
 const TEST_ADDRESS = 'regen1m0qh5y4ejkz3l5n6jlrntxcqx9r0x9xjv4vpcp';
 const TEST_BUYER_ADDRESS = 'regen13hu59094gzfcpxl58fcz294p5g5956utwlpqll';
@@ -63,7 +66,7 @@ const connect = async (): Promise<RegenApi> => {
 const runAminoTest = async (
   msgClient: MessageClient | undefined,
   testMsg: any,
-  signerAddress?: string
+  signerAddress?: string,
 ): Promise<void> => {
   let txRes: DeliverTxResponse | undefined;
 
@@ -261,8 +264,8 @@ describe('RegenApi with tendermint connection', () => {
       await runAminoTest(msgClient, TEST_MSG_CANCEL);
     });
   });
-  xdescribe('Signing and broadcasting Basket txs using legacy amino sign mode', () => {
-    it('should sign and broadcast MsgCreate', async () => {
+  describe('Signing and broadcasting Basket txs using legacy amino sign mode', () => {
+    xit('should sign and broadcast MsgCreate', async () => {
       const { msgClient } = await connect();
 
       const basketName = 'TEST' + (Date.now() % 1000);
@@ -314,7 +317,7 @@ describe('RegenApi with tendermint connection', () => {
       const { msgClient } = await connect();
       const sellOrder = {
         batchDenom: TEST_BATCH_DENOM,
-        quantity: '1',
+        quantity: '.01',
         askPrice: {
           denom: 'uregen',
           amount: '1000000',
@@ -333,7 +336,7 @@ describe('RegenApi with tendermint connection', () => {
       const { msgClient } = await connect();
       const sellOrder = {
         batchDenom: TEST_BATCH_DENOM,
-        quantity: '1',
+        quantity: '.01',
         askPrice: {
           denom: 'uregen',
           amount: '1000000',
@@ -346,7 +349,6 @@ describe('RegenApi with tendermint connection', () => {
       });
 
       await runAminoTest(msgClient, TEST_MSG_SELL);
-
     });
     xit('should sign and broadcast MsgCancelSellOrder', async () => {
       // This test passes, but it needs a real sellOrderId generated by the same address.
@@ -385,7 +387,7 @@ describe('RegenApi with tendermint connection', () => {
         orders: [
           {
             sellOrderId: '50',
-            quantity: '.01',
+            quantity: '.000001',
             bidPrice: { denom: 'uregen', amount: '1000000' },
             disableAutoRetire: false,
             retirementJurisdiction: 'US',
@@ -393,7 +395,7 @@ describe('RegenApi with tendermint connection', () => {
         ],
       });
 
-      runAminoTest(msgClient, TEST_MSG_BUY, TEST_BUYER_ADDRESS);
+      await runAminoTest(msgClient, TEST_MSG_BUY, TEST_BUYER_ADDRESS);
     });
     it('should sign and broadcast MsgBuyDirect with auto retirement disabled', async () => {
       const connectBuyer = async (): Promise<RegenApi> => {
@@ -419,36 +421,15 @@ describe('RegenApi with tendermint connection', () => {
         buyer: TEST_BUYER_ADDRESS,
         orders: [
           {
-            sellOrderId: '50',
-            quantity: '.01',
-            bidPrice: { denom: 'uregen', amount: '1000000' },
+            sellOrderId: '12', // sell order with disableAutoRetire: true
+            quantity: '.000001',
+            bidPrice: { denom: 'uregen', amount: '13000000' },
             disableAutoRetire: true,
           },
         ],
       });
 
-      runAminoTest(msgClient, TEST_MSG_BUY, TEST_BUYER_ADDRESS);
+      await runAminoTest(msgClient, TEST_MSG_BUY, TEST_BUYER_ADDRESS);
     });
   });
 });
-
-function makeEthTxHash(): string {
-  return '0x' + genRandomStr(64);
-}
-
-function makeEthContract(): string {
-  let key = crypto.randomBytes(32).toString('hex');
-  let id = '0x' + key;
-  var wallet = new ethers.Wallet(id);
-  return wallet.address;
-}
-
-function genRandomStr(length: Number): string {
-  var result = '';
-  var characters = 'abcdefghijklmnopqrstuvwxyz0123456789';
-  var charactersLength = characters.length;
-  for (var i = 0; i < length; i++) {
-    result += characters.charAt(Math.floor(Math.random() * charactersLength));
-  }
-  return result;
-}
